argos/argos_master/requirements.txt
Flask==3.0.3
flask_assets==2.1.0
flask_socketio==5.3.6
jsonschema==4.22.0
python-dotenv==1.0.1
PyYAML==6.0.1
